# docs/OpenAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DBaaS Aggregator API",
    "description": "This documentation presents the REST API for the \"Database as a Service\" (DBaaS) component. DBaaS acts as an aggregator for all adapters. It is designed to collect requests for managed databases and route them to the appropriate adapter. DBaaS stores information about all databases used in a cloud project. These databases are isolated by namespace. DBaaS uses a Classifier to identify databases within a cloud namespace. The Classifier includes service-related information such as scope, microservice name, tenant ID, and namespace.",
    "version": "main"
  },
...
  "paths": {
...
    "/api/v3/dbaas/databases/update-host": {
      "post": {
        "tags": [
          "Global Database Operation Controller v3"
        ],
        "summary": "V3. Update Physical Host in Connection Properties",
        "description": "This API updates the physical database host in the connection properties of logical databases. In the request body, you can specify whether to create a copy of the registry record (default behavior) or to make changes in place. If changes are made in place, the original database record will be updated.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UpdateHostRequest"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The host have been updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatabaseResponseV3ListCP"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Not Allowed"
          }
        },
        "security": [
          {
            "SecurityScheme": []
          }
        ]
      }
    },
    "/api/v3/dbaas/debug/internal/databases": {
      "get": {
        "tags": [
          "Controller for debug operations"
        ],
        "summary": "V3. Debug Get Logical Databases",
        "description": "Retrieves Logical Database instances in near-tabular form.\nOperation supports filters in \u0027filter\u0027 query parameter in style of RESTful Service Query Language (RSQL).",
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "description": "This parameter specifies custom RESTful Service Query Language (RSQL) query to apply filtering.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "An error occurred during getting response"
          },
          "400": {
            "description": "Incorrect RSQL query in \u0027filter\u0027 query parameter"
          },
          "200": {
            "description": "Successfully retrieved list of debug logical databases",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DebugLogicalDatabaseV3"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Not Allowed"
          }
        },
        "security": [
          {
            "SecurityScheme": []
          }
        ]
      }
    },
...
    "/api/v3/dbaas/migration/databases": {
      "put": {
        "tags": [
          "Migration controller"
        ],
        "summary": "V3. Register databases",
        "description": "This API allows you to register the database in DBaaS. The registered database would not be created, and it is assumed that it already exists in the cluster The purpose for this API is to register databases which weren not created through DBaaS. If registered database already exist in DBaaS then it will not be added.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RegisterDatabaseRequestV3"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Migration completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal error during migration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Not Allowed"
          }
        },
        "security": [
          {
            "SecurityScheme": []
          }
        ]
      }
    },
...
    "/api/v3/dbaas/{namespace}/databases": {
      "put": {
        "tags": [
          "Controller Database administration"
        ],
        "summary": "V3. Creates new database V3",
        "description": "Creates new database and returns it with connection information, or returns the already created database if it exists. This version differs from version v1 in that there is the additional require field \"dbOwner\", which should contain name of database owner. It can be for exammple microserviceName value.",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Namespace where database will be placed",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "async",
            "in": "query",
            "description": "Determines if database should be created asynchronously",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseCreateRequestV3"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "You cannot access databases in this namespace",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "There is no appropriate adapter for the specified database type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Requested role is not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Database created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Database"
                }
              }
            }
          },
          "200": {
            "description": "Already having such database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Database"
                }
              }
            }
          },
          "202": {
            "description": "Database is in process of creation"
          },
          "500": {
            "description": "Unknown error which may be related with internal work of DBaaS."
          }
        },
        "security": [
          {
            "SecurityScheme": []
          }
        ]
      }
    },
...
    "/api/v3/dbaas/{type}/physical_databases/{phydbid}/global": {
      "put": {
        "tags": [
          "Physical databases registration controller"
        ],
        "summary": "V3. Change default physical database",
        "description": "Moves the \u0027global\u0027 flag to the specified existing physical database",
        "parameters": [
          {
            "name": "phydbid",
            "in": "path",
            "description": "Physical database identifier. The value belongs to the specific database cluster",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "path",
            "description": "Type of database. Example: MongoDB, PostgreSQL, elasticsearch, etc.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated existing physical database"
          },
          "404": {
            "description": "Specified physical database does not registered"
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Not Allowed"
          }
        },
        "security": [
          {
            "SecurityScheme": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ConnectionDescription": {
        "type": "object",
        "properties": {
          "fields": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/FieldDescription"
            }
          }
        }
      },
      "Database": {
        "required": [
          "oldClassifier",
          "classifier",
          "connectionProperties",
          "namespace",
          "type",
          "adapterId",
          "name",
          "resources"
        ],
        "type": "object",
        "properties": {
          "id": {
            "description": "A unique identifier of the document in the database. This field may not be used when searching by classifier for security purpose. In appears in response when Create database API is executed.",
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/UUID"
              }
            ]
          },
          "oldClassifier": {
            "description": "Old classifier describes the purpose of the database and distinguishes this database from other database in the same namespase. It contains such keys as dbClassifier, isService, microserviceName, namespace. Setting keys depends on the database type.",
            "type": "object",
            "additionalProperties": {}
          },
          "classifier": {
            "description": "Classifier describes the purpose of the database and distinguishes this database from other database in the same namespase. It contains such keys as dbClassifier, scope, microserviceName, namespace. Setting keys depends on the database type.",
            "type": "object",
            "additionalProperties": {},
            "deprecated": true
          },
          "connectionProperties": {
            "description": "The information about connection to database. It contains such keys as url, authDbName, username, password, port, host.Setting keys depends on the database type.",
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {}
            }
          },
          "namespace": {
            "description": "Namespace where database is placed",
            "type": "string",
            "deprecated": true
          },
          "type": {
            "description": "Type of database, for example postgresql or mongodb",
            "type": "string",
            "deprecated": true
          },
          "adapterId": {
            "description": "This field indicates for which adapter the database was created.",
            "type": "string"
          },
          "name": {
            "description": "Name of database. It may be generated or, if name was specified in a request then it will be specified.",
            "type": "string"
          },
          "markedForDrop": {
            "description": "A marker indicating that the database will be deleted.",
            "type": "boolean"
          },
          "timeDbCreation": {
            "description": "Time to create a database",
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/Date"
              }
            ]
          },
          "backupDisabled": {
            "description": "This field indicates if backup is disabled or not. If true, database would not be backed up. Example: false",
            "type": "boolean"
          },
          "dbOwnerRoles": {
            "description": "The list of roles which are related to this logical database. The external security service (e.g. DBaaS Agent) can perform a verification process on this field.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "settings": {
            "description": "Additional settings for creating a database",
            "type": "object",
            "additionalProperties": {}
          },
          "connectionDescription": {
            "description": "This parameter describes connection properties.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ConnectionDescription"
              }
            ]
          },
          "warnings": {
            "description": "Lists warning messages",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "externallyManageable": {
            "description": "This parameter specifies if a control over the database is not carried out by the DBaaS adapter.",
            "type": "boolean"
          },
          "bgVersion": {
            "description": "Database version. It uses for blue-green",
            "type": "string"
          },
          "physicalDatabaseId": {
            "type": "string"
          },
          "databaseRegistry": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DatabaseRegistry"
            }
          },
          "resources": {
            "$ref": "#/components/schemas/DbResource"
          },
          "dbState": {
            "$ref": "#/components/schemas/DbState"
          }
        }
      },
...
      "DatabaseCreateRequestV3": {
        "description": "V3 Request model for adding database to DBaaS",
        "required": [
          "type",
          "originService"
        ],
        "type": "object",
        "properties": {
          "classifier": {
            "type": "object",
            "additionalProperties": {}
          },
          "type": {
            "description": "Describes the type of database in which you want to create a database. For example mongodb or postgresql",
            "type": "string"
          },
          "namePrefix": {
            "description": "This is a prefix of the database name. Prefix depends on the type of the database and it should be less than 27 characters if dbName is not specified.",
            "type": "string"
          },
          "backupDisabled": {
            "description": "This field indicates if backup is disabled or not. If true - database would not be backed up.",
            "type": "boolean"
          },
          "settings": {
            "description": "Additional settings for creating database. There is a possibility to update settings after database creation.",
            "type": "object",
            "additionalProperties": {}
          },
          "physicalDatabaseId": {
            "description": "Specifies the identificator of physical database where a logical database will be created. If it is not specified then logical database will be created in default physical database. You can get the list of all physical databases by \"List registered physical databases\" API.",
            "type": "string"
          },
          "originService": {
            "description": "Origin service which send request",
            "type": "string"
          },
          "userRole": {
            "description": "Indicates connection properties with which user role should be returned to a client",
            "type": "string"
          }
        }
      },
...
      "DatabaseResponseV3ListCP": {
        "required": [
          "classifier",
          "namespace",
          "type",
          "name",
          "connectionProperties"
        ],
        "type": "object",
        "properties": {
          "id": {
            "description": "A unique identifier of the document in the database. This field might not be used when searching by classifier for security purpose. And it exists in the response when executing Create database API",
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/UUID"
              }
            ]
          },
          "classifier": {
            "description": "Classifier describes the purpose of the database and it distinguishes this database from other databases in the same namespace. It contains such keys as dbClassifier, isServiceDb, microserviceName, namespace. Setting keys depends on the database type.",
            "type": "object",
            "additionalProperties": {}
          },
          "namespace": {
            "description": "Namespace where database is placed.",
            "type": "string"
          },
          "type": {
            "description": "Type of database, for example PostgreSQL or MongoDB",
            "type": "string"
          },
          "name": {
            "description": "Name of database. It may be generated or, if name was specified in the request, then it will be specified.",
            "type": "string"
          },
          "externallyManageable": {
            "description": "This parameter specifies if a control over the database is not carried out by the DBaaS adapter",
            "type": "boolean"
          },
          "timeDbCreation": {
            "description": "Time to create a database.",
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/Date"
              }
            ]
          },
          "settings": {
            "description": "Additional settings for creating a database.",
            "type": "object",
            "additionalProperties": {}
          },
          "backupDisabled": {
            "description": "This field indicates if backup is disabled or not. If true, database would not be backed up. Example: false",
            "type": "boolean"
          },
          "physicalDatabaseId": {
            "description": "Physical database identifier where the registered database should be located. If it is absent, adapter id may be used to identify the target physical database.",
            "type": "string"
          },
          "connectionProperties": {
            "description": "This is an information about connection to database. It contains such keys as url, authDbName, username, password, port, host.Setting keys depends on the database type.",
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {}
            }
          },
          "resources": {
            "$ref": "#/components/schemas/DbResource"
          }
        }
      },
      "DatabaseStateStatus": {
        "enum": [
          "PROCESSING",
          "CREATED",
          "DELETING",
          "DELETING_FAILED",
          "ARCHIVED",
          "ORPHAN"
        ],
        "type": "string"
      },
...
      "Date": {
        "format": "date",
        "type": "string",
        "example": "2022-03-10"
      },
      "DbResource": {
        "required": [
          "kind",
          "name"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "description": "The kind of resource. For example database or user",
            "type": "string"
          },
          "name": {
            "description": "Name of the resource.",
            "type": "string"
          }
        }
      },
      "DbState": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "state": {
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/DatabaseStateStatus"
              }
            ],
            "deprecated": true
          },
          "databaseState": {
            "$ref": "#/components/schemas/DatabaseStateStatus"
          },
          "description": {
            "type": "string"
          },
          "podName": {
            "type": "string"
          }
        }
      },
      "DebugDatabaseDeclarativeConfigV3": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "settings": {
            "type": "object",
            "additionalProperties": {}
          },
          "lazy": {
            "type": "boolean"
          },
          "instantiationApproach": {
            "type": "string"
          },
          "versioningApproach": {
            "type": "string"
          },
          "versioningType": {
            "type": "string"
          },
          "classifier": {
            "type": "object",
            "additionalProperties": {}
          },
          "type": {
            "type": "string"
          },
          "namePrefix": {
            "type": "string"
          },
          "namespace": {
            "type": "string"
          }
        }
      },
      "DebugLogicalDatabaseV3": {
        "type": "object",
        "properties": {
          "namespace": {
            "type": "string"
          },
          "microservice": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "logicalDbName": {
            "type": "string"
          },
          "bgVersion": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "physicalDbId": {
            "type": "string"
          },
          "physicalDbAdapterUrl": {
            "type": "string"
          },
          "declaration": {
            "$ref": "#/components/schemas/DebugDatabaseDeclarativeConfigV3"
          }
        }
      },
...
      "FieldDescription": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/FieldTypeEnum"
          }
        }
      },
      "FieldTypeEnum": {
        "enum": [
          "PASSWORD"
        ],
        "type": "string"
      },
...
      "RegisterDatabaseRequestV3": {
        "description": "Request to add database to registration",
        "required": [
          "classifier",
          "connectionProperties",
          "resources",
          "namespace",
          "type",
          "name"
        ],
        "type": "object",
        "properties": {
          "classifier": {
            "description": "Classifier describes the purpose of database, it should not exist in registration, or the request would be ignored.",
            "type": "object",
            "additionalProperties": {}
          },
          "connectionProperties": {
            "description": "Connection properties used to connect to database. It contains such keys as url, authDbName, username, password, port, host. Setting keys depends on the database type.",
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {}
            }
          },
          "resources": {
            "description": "List of the resources needed to drop during database drop.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DbResource"
            }
          },
          "namespace": {
            "description": "Namespace where database is placed.",
            "type": "string"
          },
          "type": {
            "description": "The type of database, for example postgresql or mongodb.",
            "type": "string"
          },
          "adapterId": {
            "description": "Identifier of an adapter to work with database. If not specified then the default would be used.",
            "type": "string"
          },
          "name": {
            "description": "Name of database.",
            "type": "string"
          },
          "backupDisabled": {
            "description": "This parameter specifies if the DBaaS should except this database from backup/restore procedure. The parameter cannot be modified and it is installed only once during registration request.",
            "type": "boolean"
          },
          "physicalDatabaseId": {
            "description": "Physical database identifier where the registered database should be located. If it is absent, adapter id may be used to identify the target physical database.",
            "type": "string"
          },
          "dbHost": {
            "description": "Physical database host where the registered database is located. Must be in format: \u003cservice-name\u003e.\u003cnamespace\u003e, e.g.: pg-patroni.postgresql-core",
            "type": "string"
          }
        }
      },
...
      "UUID": {
        "format": "uuid",
        "pattern": "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
        "type": "string"
      },
...
      "UpdateHostRequest": {
        "description": "Request object containing the necessary information to update the physical host in a logical database",
        "required": [
          "type",
          "classifier",
          "physicalDatabaseHost",
          "physicalDatabaseId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "description": "The physical type of logical database. For example mongodb or postgresql",
            "type": "string"
          },
          "classifier": {
            "description": "The unique key of existing database. ",
            "type": "object",
            "additionalProperties": {}
          },
          "makeCopy": {
            "description": "If true, a copy of the logical database registry will be created before applying changes. If false, changes will be made directly to the existing registry.",
            "default": true,
            "type": "boolean"
          },
          "physicalDatabaseHost": {
            "description": "Contains physical database host on which it\u0027s needed to change. The host must follow the format: \u003cphysical database k8s service\u003e.\u003cphysical database namespace\u003e. Example: pg-patroni.core-postgresql.",
            "type": "string"
          },
          "physicalDatabaseId": {
            "description": "Specifies id of new physical database. The value can be found in dbaas-adapter env",
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "SecurityScheme": {
        "type": "http",
        "description": "Authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}